+++ agent/api.go
@@ -34,14 +34,14 @@ func newServer(a *Agent) *http.Server {
 
 func registerRoutes(a *Agent, r *mux.Router) {
 	r.HandleFunc("/events", addEvent(a)).Methods(http.MethodPost)
-	r.HandleFunc("/healthz", healthz(a.conn)).Methods(http.MethodGet)
+	r.HandleFunc("/healthz", healthz(a.Connected)).Methods(http.MethodGet)
 }
 
 // healthz returns an OK status if the agent is up and connected to a backend.
 // If the backend connection is closed, it returns service unavailable.
-func healthz(conn transport.Transport) http.HandlerFunc {
+func healthz(connected func() bool) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		if conn.Closed() {
+		if !connected() {
 			w.WriteHeader(http.StatusServiceUnavailable)
 			_, _ = fmt.Fprint(w, "sensu backend unavailable")
 			return